+++ components/common/CharmEditor/components/nft/nftUtils.ts
@@ -1,24 +1,7 @@
-import type { NodeAttrs } from './nftSpec';
+import type { NftNodeAttrs } from 'lib/nft/interface';
 
 // a function to extract user screen name and tweet id from a tweet url
-export function extractAttrsFromUrl(url: string): NodeAttrs | null {
-  if (!url) {
-    return null;
-  }
-
-  const match = url.match(/opensea\.io\/([^/]+\/)?assets\/ethereum\/([^/]+)\/([^/]+)/);
-  if (!match) {
-    return null;
-  }
-  return {
-    chain: 1,
-    contract: match[2],
-    token: match[3]
-  };
-}
-
-// a function to extract user screen name and tweet id from a tweet url
-export function extractAttrsFromEmbedCode(url: string): NodeAttrs | null {
+export function extractAttrsFromEmbedCode(url: string): NftNodeAttrs | null {
   if (!url) {
     return null;
   }
