+++ lib/nft/interface.ts
@@ -1,0 +1,7 @@
+import type { supportedChainIds } from 'connectors';
+
+export type NftNodeAttrs = {
+  chain: typeof supportedChainIds[number];
+  contract: string;
+  token: string;
+};
