+++ lib/twitter/interface.ts
@@ -1,0 +1,4 @@
+export type TweetNodeAttrs = {
+  id: string;
+  screenName: string;
+};
